Feature/UI modernization v2

+++ tracker/appTracker/views.py
@@ -1,3 +1,5 @@
+from urllib import request
+
 from django.shortcuts import render, redirect, get_object_or_404
 
 from django.contrib.auth.models import User
@@ -6,6 +8,8 @@ from django.contrib.auth.decorators import login_required
 from .models import Expense, Income
 from django.db.models import Sum
 from django.db.models.functions import TruncMonth
+from django.contrib import messages
+from appTracker.models import Expense, Income
 
 
 # Create your views here.
@@ -25,13 +29,13 @@ def login_view(request):
         
         if user is not None:
             login(request, user)
-            return redirect('dashboard') #this is key
+            return redirect('manage') #this is key
         else:
            return render(request,
-                          'login.html',
+                          'accounts/login.html',
                           {'error': 'Invalid username or password'}
                           )
-    return render(request, 'login.html')
+    return render(request, 'accounts/login.html')
 
 def registration(request):
     if request.method == "POST":
@@ -41,41 +45,102 @@ def registration(request):
 
         # Check if user already exists
         if User.objects.filter(username=username).exists():
-            return render(request, 'reg.html', {'error': 'Username already exists'})
+            return render(request, 'accounts/register.html', {'error': 'Username already exists'})
 
         user = User.objects.create_user(username=username, email = email, password=password )
         user.save()
         
         return redirect('login')
-    return render(request, 'reg.html')  
+    return render(request, 'accounts/register.html')  
 
 @login_required
 def manage_expenses(request):
 
     if request.method == "POST":
-        if 'income_submit' in request.POST: 
+        if 'income_submit' in request.POST:
             amount = request.POST.get('income')
-            Income.objects.create(user = request.user, amount=amount)
+            Income.objects.create(user=request.user, amount=amount)
+            messages.success(request, "Income added successfully!")
+
         else:
-            title=request.POST.get('title')
-            amount=request.POST.get('amount')
-            category=request.POST.get('category')
-            date=request.POST.get('date')
+            title = request.POST.get('title')
+            amount = request.POST.get('amount')
+            category = request.POST.get('category')
+            date = request.POST.get('date')
 
-            Expense.objects.create(user = request.user, title=title, amount=amount, category=category, date=date)
+            Expense.objects.create(
+                user=request.user,
+                title=title,
+                amount=amount,
+                category=category,
+                date=date
+            )
+
+            messages.success(request, "Expense added successfully!")
+
         return redirect('manage')
-    # only logged-in user's data'
-    expenses = Expense.objects.filter(user = request.user)
-    incomes = Income.objects.filter(user = request.user)
 
-    total_income = sum([i.amount for i in incomes])
+    # Fetch data
+    expenses = Expense.objects.filter(user=request.user)
 
-    return render(request, 'manage.html', {'expenses':expenses, 'total_income': total_income})
+    # MONTHLY SUMMARY
+    monthly_queryset = (
+    expenses
+    .annotate(month=TruncMonth("date"))
+    .values("month")
+    .annotate(total=Sum("amount"))
+    .order_by("month")
+)
+    monthly_data = []
+    
+    for item in monthly_queryset:
+        monthly_data.append({
+                "month": item["month"].strftime("%b %Y"),
+                "total": float(item["total"])
+                })
+
+    incomes = Income.objects.filter(user=request.user)
+
+    total_income = incomes.aggregate(total=Sum("amount"))["total"] or 0
+
+    total_expense = expenses.aggregate(total=Sum("amount"))["total"] or 0
+
+    balance = total_income - total_expense
+
+    transaction_count = expenses.count()
+
+    category_queryset = (
+    expenses
+    .values("category")
+    .annotate(total=Sum("amount"))
+    .order_by("category")
+    )
+    
+    category_data = []
+    
+    for item in category_queryset:
+        category_data.append({
+            "category": item["category"],
+            "total": float(item["total"])
+        })
+
+    context = {
+        "expenses": expenses,
+        "total_income": total_income,
+        "total_expense": total_expense,
+        "balance": balance,
+        "transaction_count": transaction_count,
+        "monthly_data": monthly_data,
+        "category_data": category_data,
+    }
+
+    return render(request, "dashboard/manage.html", context)
 
 # edit expense
 
+@login_required
 def edit_expense(request, id):
-    expense =  get_object_or_404(Expense, id=id)
+    expense =  get_object_or_404(Expense, id=id, user=request.user)
 
     if request.method=="POST":
         expense.title = request.POST.get('title')
@@ -84,19 +149,26 @@ def edit_expense(request, id):
         expense.date = request.POST.get('date')
 
         expense.save()
+        messages.success(request, "Expense updated successfully.")
 
         return redirect('manage')
-    return render(request,'edit_expense.html', {'expense' : expense})
+    return render(request,'expenses/edit_expense.html', {'expense' : expense})
 
 # Delete Expense
 
+@login_required
 def delete_expense(request, id):
 
-    expense = get_object_or_404(Expense, id=id)
+    if request.method == "POST":
+        expense = get_object_or_404(
+            Expense,
+            id=id,
+            user=request.user
+        )
+        expense.delete()
+        messages.success(request, "Expense deleted successfully.")
 
-    expense.delete()
-
-    return redirect('manage')
+    return redirect("manage")
     
 # Data_Visualization
 @login_required
@@ -115,28 +187,8 @@ def data_visualization(request):
 
     if end_date: expenses = expenses.filter(date__lte=end_date)
 
-    # MONTHLY SUMMARY
-    monthly_data = (expenses.annotate(month=TruncMonth('date'))
-                    .values('month')
-                    .annotate(total=Sum('amount'))
-                    .order_by('month'))
-
-    # TOTAL CALCULATIONS
-    total_expense = expenses.aggregate(total=Sum('amount'))['total'] or 0
-
-    total_income = Income.objects.filter(user = request.user)\
-        .aggregate(total=Sum('amount'))['total'] or 0
-
-    balance = total_income - total_expense
-
-    context = {
-        'expenses':expenses,
-        'monthly_data':monthly_data,
-        'total_expense':total_expense,
-        'total_income': total_income,
-        'balance': balance,
-    }
-    return render(request,'dataVisualization.html', context)
+    
+    return render(request,'dashboard/visualization.html', context)
 
 # LOGOUT FUNCTION
 def logout_view(request):
